Generate moves and Board changes
Board can now move Pieces. Plan to add ways of checking if Pieces can be moved in move Piece. May use another function for this. May add functionality for updating check and checkmate here.

Spot now changes the name of its piece when the Piece is moved.

Chess now launches the game.

Game now has a way of taking inputs from users to move pieces. Still needs a method for checking Stalemates and Checkmates

King and Knight now generate moves in a Set for O(1) access.

Bishops can check if a move is valid. Needs to be able to generate list of moves.

Pawn no longer changes if it has moved. Has moved state set by board.

Piece abstract class now has a place for the move list for each piece.

Rook now has a method of checking if a move is legal. Still needs to be tested. Also Needs to generate a list of moves.

Board testing is being used for a lot of testing will update to only do proper testing if time is available.