Resetting how canMove will work
Pieces now require their current row and col for canMove so set up that board will give it to them on initialization.

Piece will not have a way of seeing if a move is valid, meaning a spot is open or has a caputre.

Piece will also be able to tell if the target spot is the same spot. May go away from this.

King is first to use new canMove direction.