Checkmate working
Added is checkmate to board to check if checkmate.
Changed isIllegal to use a temp set of pieces to avoid ConcurrentModificationException.

All pieces now check that end spot is not empty or hold same color piece.

Game now checks player input and checks for incheck and checkmate.

Plan to set up Stalemate and clean up next.